cls_item: fix suffix parsing
--breaking change--

First of all determine the override delimiter automatically, so
either : (for the new syntax) or _ (for the old syntax) is used.

Change the way override are detected, now splitting them off
properly from the variable or function name.

Introduce new Attributes VarNameComplete and FuncNameComplete
which reflect the complete variable or function name
including all overrides and flags

Remove support for PkgSpec as this attribute has no real use
as everything is treated as an override

Closes #18
Closes #17