stash: allow filtering by multiple elements
classifier, attribute and attributeValue now also
take lists, sets or tuples as arguments and filter the stash
by those settings.
Filtering is done by a logical OR, so the item in question
needs to match any of the setup criteria.
Logical AND filtering needs to be still done calling GetItemsFor
multiple times and doing the computation outside of the stash.

Closes #142

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>