stash: fix GetLoneAppends
if a bbappend would have contained a require or
include statement, the links of the include file
would have been computed the same as a bb file,
thus preventing to identify the bbappend as a
scanable file, if it would have been passed alone
into the stash parsing.

Fix that by limiting the "linked appends" to bb files only.

Closes #150

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>