cls_item: deprecate Links, IncludedFrom and AddLink
as the information is not correct within multiple corner cases
and could lead to wrong assumptions.
By default the Stash returns the correct set of files, so
there is no need to expose this info anymore.

AddLink was removed entirely, as it shouldn't be used from
the outside anyway.
Links and IncludedFrom are marked deprecated and return
and empty set now.

Relates to #201

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>