stash: add fingerprint
which gives a checksum of the current Stash object
including the parser relevant options and all of the
extracted objects.
This can help to decide if there has been changes to
a known baseline.
Those baselines can only be safely calculated on the
very same host from the same setup, as the Stash objects
contain absolute paths.

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>