data: cleanup standard data
remove all data which is not a 100% mandatory for the
parser to work.
Everything else needs to be moved into a consuming
caller's codebase.
Like this we can make sure everything stays up to date
and can e.g. made release specific (esp with the ever
changing amount of functions, machines and variables,
this was an long overdue move)

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>