parser: fix function block handling
when the end of a function or python block is detected,
we need to return the parsed block so far and on the next
cycle create the next block.
In before that was not correct, as e.g. a python block
after a python block was put before the 1st python block,
which created incorrect Line attributes and a lot more
issues.
Also the scope level calculation in for function blocks
was erroneous.

A few more tests were added to avoid future regressions

Fixes https://github.com/priv-kweihmann/oelint-adv/issues/771

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>